+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.19.1
+
+### Patch Changes
+
+-   91e95ef8: BUG - Generated Controller Extensions do not always work with TS
+-   Updated dependencies [91e95ef8]
+    -   @sap-ux/ui5-application-writer@0.22.1
+
+## 0.19.0
+
+### Minor Changes
+
+-   7ac0cb40: Add new module containing an eslint plugin with reusable eslint configs
+
+### Patch Changes
+
+-   Updated dependencies [7ac0cb40]
+    -   @sap-ux/ui5-application-writer@0.22.0
+
 ## 0.18.1
 
 ### Patch Changes