+++ packages/ui5-application-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui5-application-writer
 
+## 0.22.1
+
+### Patch Changes
+
+-   91e95ef8: BUG - Generated Controller Extensions do not always work with TS
+
+## 0.22.0
+
+### Minor Changes
+
+-   7ac0cb40: Add new module containing an eslint plugin with reusable eslint configs
+
 ## 0.21.1
 
 ### Patch Changes